+++ NewLife.CubeNC/ViewModels/SearchField.cs
@@ -1,4 +1,6 @@
-﻿namespace NewLife.Cube.ViewModels;
+﻿using System.Text.Json.Serialization;
+
+namespace NewLife.Cube.ViewModels;
 
 /// <summary>搜索字段</summary>
 public class SearchField : DataField
@@ -8,10 +10,12 @@ public class SearchField : DataField
     /// 多选框返回的数据有多条，需要逗号分隔；
     /// 如果没有选中任何项，则没有返回，此时会强制覆盖Url参数中的同名字段，避免取消选中无效的问题。
     /// </remarks>
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     public Boolean Multiple { get; set; }
 
 #if MVC
     /// <summary>视图。MVC特有，允许针对字段定义视图</summary>
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     public String View { get; set; }
 #endif
 }